Bot/requirements.txt
discord.py
python-dotenv
discord-py-interactions